+++ test/host/test_planner_math.cpp
@@ -71,12 +71,12 @@ int main(void) {
     float v1 = 50.0f;
     float a = 200.0f;
     float step_hz = 16000.0f; /* above PLANNER_PACK_MIN_HZ */
-    float dt1 = 1.0f / step_hz;
-    float dtn = 64.0f / step_hz;
-    float phi1 = planner_sine_advance_phi(0.0f, v0, v1, a, dt1);
-    float phin = planner_sine_advance_phi(0.0f, v0, v1, a, dtn);
-    expect_true("pack-dt n advances more than 1", phin > phi1 * 10.0f);
-    expect_true("pack-dt n not instantly done", phin < 1.0f);
+    double dt1 = 1.0 / (double)step_hz;
+    double dtn = 64.0 / (double)step_hz;
+    double phi1 = planner_sine_advance_phi(0.0, v0, v1, a, dt1);
+    double phin = planner_sine_advance_phi(0.0, v0, v1, a, dtn);
+    expect_true("pack-dt n advances more than 1", phin > phi1 * 10.0);
+    expect_true("pack-dt n not instantly done", phin < 1.0);
   }
 
   /* reverse_decel helpers */
